fix: simplified theme storage
