+++ packages/frontend/src/common/components/buttons/ToggleThemeButton.tsx
@@ -2,18 +2,15 @@ import React, { useContext } from "react";
 import { Brightness6 } from "@mui/icons-material";
 import { Button, Typography } from "@mui/material";
 import { ThemeContext } from "../../context/ThemeContext";
-import { LocalStorage, ThemeStatus } from "../../utils/localStorage";
 
 export function ToggleThemeButton() {
-  const { currentTheme, setDarkTheme, setLightTheme } = useContext(ThemeContext);
+  const { currentTheme, setTheme } = useContext(ThemeContext);
 
   function toggleTheme() {
     if (currentTheme.palette.mode === "dark") {
-      LocalStorage.setThemePreference(ThemeStatus.light);
-      setLightTheme();
+      setTheme("light");
     } else {
-      LocalStorage.setThemePreference(ThemeStatus.dark);
-      setDarkTheme();
+      setTheme("dark");
     }
   }
