+++ packages/frontend/src/common/context/ThemeContext.tsx
@@ -1,135 +1,41 @@
 import React, { PropsWithChildren, useMemo, useState } from "react";
-import {
-  createTheme,
-  responsiveFontSizes,
-  Theme,
-  ThemeOptions,
-  useMediaQuery,
-} from "@mui/material";
+import { responsiveFontSizes, Theme, useMediaQuery } from "@mui/material";
 import { ThemeProvider } from "@mui/material/styles";
 import { useLocalStorage } from "../hooks/useLocalStorage";
-import { LocalStorage, ThemeStatus } from "../utils/localStorage";
+import { LocalStorage } from "../utils/localStorage";
+import { RetroPaletteMode } from "../../../mui.types";
+import { themes } from "./themes";
 
 export interface ThemeContextValues {
   currentTheme: Theme;
-  setDarkTheme: () => void;
-  setLightTheme: () => void;
+  setTheme: (theme: RetroPaletteMode) => void;
 }
-
-const components: ThemeOptions["components"] = {
-  MuiTypography: {
-    defaultProps: {
-      variantMapping: {
-        title: "h1",
-        menuTitle: "span",
-        label: "div",
-      },
-    },
-  },
-};
-const typography: ThemeOptions["typography"] = {
-  htmlFontSize: 16,
-  h1: {
-    fontSize: "4rem",
-  },
-  h2: {
-    fontSize: "3rem",
-  },
-  h3: {
-    fontSize: "2rem",
-  },
-  h4: {
-    fontSize: "1.8rem",
-  },
-  title: {
-    fontSize: "2.125rem",
-    fontWeight: "400",
-  },
-  menuTitle: {
-    fontSize: "2.125rem",
-    fontFamily: "Permanent Marker, cursive",
-  },
-  label: {
-    fontSize: "1.25rem",
-  },
-};
-const lightTheme = createTheme({
-  palette: {
-    mode: "light",
-    primary: {
-      main: "#303030",
-    },
-    secondary: {
-      main: "#00b4f0",
-    },
-    background: {
-      default: "#B0BEC5",
-      paper: "#CFD8DC",
-    },
-    highlightColorPrimary: "#f05a96",
-    highlightColorSecondary: "#82c864",
-  },
-  ...typography,
-  ...components,
-});
-
-const darkTheme = createTheme({
-  palette: {
-    mode: "dark",
-    primary: {
-      main: "#EFEFEF",
-    },
-    secondary: {
-      main: "#b41964",
-    },
-    background: {
-      default: "#191c38",
-      paper: "#191c38",
-    },
-    highlightColorPrimary: "#b41964",
-    highlightColorSecondary: "#14b400",
-  },
-  typography: {
-    ...typography,
-  },
-  components: {
-    ...components,
-  },
-});
 
 export const ThemeContext = React.createContext<ThemeContextValues>(undefined!);
 
 export function ThemeContextProvider(props: PropsWithChildren) {
   const prefersDarkMode = useMediaQuery("(prefers-color-scheme: dark)");
-  const [themeChoice, setThemeChoice] = useState<ThemeStatus>(ThemeStatus.not_set);
+  const [themeChoice, setThemeChoice] = useState<RetroPaletteMode | undefined>(undefined);
 
-  const currentTheme = useMemo(() => {
-    switch (themeChoice) {
-      case ThemeStatus.dark:
-        return darkTheme;
-      case ThemeStatus.light:
-        return lightTheme;
-      case ThemeStatus.not_set:
-        return prefersDarkMode ? darkTheme : lightTheme;
+  const currentTheme: Theme = useMemo(() => {
+    if (themeChoice === undefined || themes[themeChoice] === undefined) {
+      return prefersDarkMode ? (themes["dark"] as Theme) : (themes["light"] as Theme);
     }
+    return themes[themeChoice] as Theme;
   }, [prefersDarkMode, themeChoice]);
 
   useLocalStorage(() => {
     setThemeChoice(LocalStorage.getThemePreference());
   });
 
-  function setDarkTheme() {
-    setThemeChoice(ThemeStatus.dark);
-  }
-
-  function setLightTheme() {
-    setThemeChoice(ThemeStatus.light);
+  function setTheme(theme: RetroPaletteMode) {
+    setThemeChoice(theme);
+    LocalStorage.setThemePreference(theme);
   }
 
   const contextValues = {
     currentTheme,
-    setDarkTheme,
-    setLightTheme,
+    setTheme,
   };
 
   return (
